I/O in rv.ipynb changed to use tests module
